Make sure that headers scrolled in view on mobile
Currently, when clicking on a ToC link in a single page web version, the title for that section is hidden under the fixed banner. We can use `scroll-margin-top` to add a little additional scroll amount to scroll the header into view. `3em` matches the current top margin of `3em` on headings and hgroups.