Validate that strings aren't too long
Without this check, we can get errors of this form:

`Uncaught PDOException: SQLSTATE[22001]: String data, right truncated: 1406 Data too long for column`

Most MySQL DB columns are varchar(255), but artist name is 191, which is the
longest it could be an still be part of the unique key index. That is why 191
is set as the constant.