Require UrlName to be unique for Artist+Artwork
Before this change, `Artwork::Validate()` would allow multiple `unverified`
submissions with the same Artist UrlName and Artwork UrlName, and the moderator
could approve only one.

Now, only one submission is allowed with a Artist+Artwork combo of UrlNames.
When a user gets a validation error on the second submission, it hints to them
that the same artworks is already in the queue and doesn't need to be submitted
again.